requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0

# Database & utilities  
Flask-SQLAlchemy==3.0.5

# Optional: Face Recognition (comment out for simple deployment)
# opencv-python==4.8.1.78
# face-recognition==1.3.0  
# Pillow==10.0.1
# numpy==1.24.4
# redis==5.0.1
# python-dateutil==2.8.2
